requirements.txt
Pyrebase==3.0.27
numpy==1.19.3
python-dotenv==0.15.0
pillow==8.0.1
opencv-python
